fix: address PR #1071 review — panel split, imports, pending PR map
