+++ apps/web/components/vcs/vcs-dialogs.tsx
@@ -28,7 +28,7 @@ import {
 import { useSessionGit } from "@/hooks/domains/session/use-session-git";
 import { useRepoDisplayName } from "@/hooks/domains/session/use-repo-display-name";
 import { useGitOperations } from "@/hooks/use-git-operations";
-import { useAppStore } from "@/lib/state/store";
+import { useAppStore } from "@/components/state-provider";
 import { useGitWithFeedback } from "@/hooks/use-git-with-feedback";
 import { useUtilityAgentGenerator } from "@/hooks/use-utility-agent-generator";
 import { useIsUtilityConfigured } from "@/hooks/use-is-utility-configured";
